backend/requirements.txt
# FastAPI
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Task Queue
celery==5.4.0
redis==5.2.1
flower==2.0.1

# AI
google-generativeai==0.8.3
boto3==1.35.86

# Web Scraping
playwright==1.49.1
beautifulsoup4==4.12.3
lxml==5.3.0
httpx==0.28.1

# Storage (S3/R2)
boto3==1.35.86
botocore==1.35.86

# Auth
propelauth-fastapi==2.3.0

# Config
pydantic-settings==2.7.0
python-dotenv==1.0.1

# Utils
Pillow==11.1.0
python-magic==0.4.27
aiofiles==24.1.0
orjson==3.10.13
tenacity==9.0.0

# Email
sendgrid==6.11.0

# Monitoring
posthog==3.8.3
